core, eth, les: implement unclean-shutdown marker (#21893)
This PR implements unclean shutdown marker. Every time geth boots, it adds a timestamp to a list of timestamps in the database. This list is capped at 10. At a clean shutdown, the timestamp is removed again.
Thus, when geth exits unclean, the marker remains, and at boot up we show the most recent unclean shutdowns to the user, which makes it easier to diagnose root-causes to certain problems.

Co-authored-by: Nagy Salem <me@muhnagy.com>
# Conflicts:
#	core/rawdb/accessors_metadata.go
#	core/rawdb/database.go
#	core/rawdb/schema.go
#	eth/backend.go
#	les/client.go