Docker: explicitly declare base image source
Some container engines such as podman have started to prompt the user to explicitly state (thus sometimes causing service scripts on systemd) to fail for lack of mentioning what container image hub is used. This change declares the hub to be docker.io, thus removing the prompt on systems that use podman.