update libmdbx submodule to v0.11.1.1.
In libmdbx >= v0.11.1, the database signature is automatically updated,
which is important to prevent false positive diagnostics of a database
corruption due to the shortcomings of older versions of the library.

Please refer to the ChangeLog for more historical information and actual TODOs.
https://github.com/erthink/libmdbx/blob/master/ChangeLog.md

Release notes: https://github.com/erthink/libmdbx/releases/tag/v0.11.1

Signed-off-by: Leonid Yuriev <leo@yuriev.ru>