Update README.md
Updated the "getting started" instructions to direct normal users to building the latest alpha release (rather than devel).

A similar change was previously made to the README.md in the alpha branch only. However, since the git repo defaults to the devel branch, most users will still see the README.md in the devel branch first. This may inadvertently guide new users to build the devel branch. Hence the change to the README.md in the devel branch.

An alternative option would be to default the git repo to the alpha branch, rather than the devel branch.