Enhance GetProof to support historical proofs
When the block number is older than the current head, GetProof will
create a memdb instance to temporarily rewind the hashed state table.
Then, it builds a retain list to invalidate the trie nodes which differ
from the current trie root.  To do each of these, it leverages the
existing stage unwinding code (with a few wiring changes).

When the block number is too old, this unwinding could become
computationally intensive (we know the upper limit to re-compute the
hash trie from scratch is currently around 15 minutes on mainnet).  So,
this change introduces the somewhat arbitrarily chosen limit of 1000
blocks to allow rewinding.  If this is too much or too little in
practice, this limit can be revised, or made configurable.