Fix bugs in trie hash computation
There are currently a number of bugs in the flat DB trie hash
computation.  These bugs are improbable in 'real' scenarios (hence the
ability to sync the assorted large chains), and, the repercussions of
encountering them are low (generally re-computing the hash at a later
block will succeed).  Still, they can cause the process to crash, or
deadlock, and a clever adversary could feasibly construct a block or
storage update designed to trigger these bugs. (Or, an unlucky block may
trigger them inadvertently). Based on the tracing in the code, it seems
that some of these bugs may have been witnessed in the wild, but not
reliably reproduced (for instance when `maxlen >= len(curr)`).

1. There is an infinite loop that can occur in
   _nextSiblingOfParentInMem.  This occurs in the account path when the
   c.k[1] entry is 'nil' and the code will continuously retry resolving
   this entry indefinitely.

2. When the next trie node is deeper in the trie than the previous, but
   is not a descendent of the previous trie node, then any higher trie
   nodes which are not overwritten are inadvertently left in memory
   (when they should be nil-ed) resulting in an incorrect hash or index
   out of bounds panic.

3. When the last trie node being processed is comprised entirely of 'f'
   nibbles, the 'FirstNotCoveredPrefix' returns an empty byte array
   because there is no next nibble sub-tree.  This causes keys to
   inappropriately be reprocessed triggering either an index out of
   bounds panic or incorrect hash results.

4. When the _nextSiblingInDB path is triggered, if the next nibble
   subtree contains no trie table entries, then any keys with a prefix
   in that subtree will be skipped resulting in incorrect hash results.

The fuzzing seeds included cover all four of these cases (for both
accounts and storage where appropriate).  As fuzzing is baked into the
native go 1.18 toolchain, running 'go test' (as the Makefile currently
does) is adequate to cover these failures.  To run additional fuzzing
the `-fuzz` flag may be provided as documented in the go test doc.