Enable negative Merkle proofs
This addresses the last known deficiency of the eth_getProof
implementation.  The previous code would return an error in the event
that the element was not found in the trie.  EIP-1186 allows for
'negative' proofs where a proof demonstrates that an element cannot be
in the trie, so this commit updates the logic to support that case.