tests/bor, consensus/bor :: add : TestMiningBenchmark, e2e testHelpers, GetCurrentValidatorsFromContract (#8060)
In this PR, we setup e2e testing capabilities in Erigon. With e2e
capabilities, a unit-case will be able to spawn multiple-node network
with the same validators as extracted from the genesis validatorSet.
This is a foundation for a testing suite which will be capable to test
hardforks, reorgs, critical features in just a unit test. The genesis
test of this suite is TestMiningBenchmark.

`CGO_CFLAGS="-D__BLST_PORTABLE__" go test -run ^TestMiningBenchmark$
github.com/ledgerwatch/erigon/tests/bor -v -count=1` can be used to run
the test.

Duplicate of #7998 .