eth: do not reply with empty block headers
This is a hack to make us work with erigon 2.48 peers that have --sentry.drop-useless-peers
If we reply with an empty list, we're going to be considered useless and kicked.
Once enough of erigon nodes are updated in the network past this commit, this check should be removed,
because it is totally acceptable to return an empty list.